requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.34
openpyxl==3.1.5
reportlab==4.2.2
python-dotenv

